add custom hook for signup, and fix gender issue
