+++ src/components/form/Select/index.js
@@ -1,20 +1,27 @@
 import './styles.scss';
 
-const Select = ({ name, options, placeholder, register, handleFocus }) => {
+const Select = ({ name, options, placeholder, register, handleFocus, error }) => {
   return (
     <div className="select">
-      <select id={name} name={name} {...register(name)} onFocus={handleFocus}>
+      <select
+        id={name}
+        name={name}
+        {...register(name)}
+        onFocus={handleFocus}
+        defaultValue={placeholder}
+      >
         {placeholder && (
-          <option disabled selected value="">
+          <option disabled value={placeholder}>
             {placeholder}
           </option>
         )}
         {options.map((opt, index) => (
           <option key={`select-option-${index}`} value={opt}>
-            {opt}
+            {opt.toUpperCase()}
           </option>
         ))}
       </select>
+      <small className="error-message">{error?.message}</small>
     </div>
   );
 };
