+++ src/components/signup/SignupForm/index.js
@@ -1,63 +1,18 @@
-import { useCallback, useEffect } from 'react';
-import { useDispatch } from 'react-redux';
-import { useForm } from 'react-hook-form';
-import { zodResolver } from '@hookform/resolvers/zod';
-import { Redirect, useHistory } from 'react-router-dom';
-import { z } from 'zod';
+import { Redirect } from 'react-router-dom';
 import Input from 'components/form/Input';
 import Button from 'components/common/Button';
 import routesPaths from 'routes/routesPaths';
 import useTranslation from 'hooks/useTranslation';
-import useAuth from 'hooks/useAuth';
-import { api } from 'services/api';
-import { useSignupMutation } from 'services/auth/auth';
-import { setLoggedInUser } from 'utils/auth';
-import { PASSWORD_REGEX } from 'constants/constants';
 import Select from 'components/form/Select';
 import './styles.scss';
+import { useSignup } from 'hooks/useSignup';
 
 const SignupForm = () => {
   const t = useTranslation();
-  const dispatch = useDispatch();
-  const { push } = useHistory();
-  const { user, authenticated } = useAuth();
-  const [signup, { isLoading, isSuccess, error }] = useSignupMutation();
-
-  const schema = z
-    .object({
-      name: z.string(),
-      gender: z.string(),
-      email: z.string().email({ message: t('signup.errors.emailMsg') }),
-      password: z.string().regex(PASSWORD_REGEX, { message: t('signup.errors.passwordMsg') }),
-      passwordConfirmation: z
-        .string()
-        .regex(PASSWORD_REGEX, { message: t('signup.errors.passwordMsg') }),
-    })
-    .refine(data => data.password === data.passwordConfirmation, {
-      message: t('signup.errors.passwordConfirmationMsg'),
-      path: ['passwordConfirmation'],
-    });
-
-  const {
-    register,
-    handleSubmit,
-    formState: { errors },
-  } = useForm({ resolver: zodResolver(schema) });
-
-  const onSubmit = data => signup(data);
-
-  const resetErrors = useCallback(() => dispatch(api.util.resetApiState()), [dispatch]);
-
-  const handleFocus = () => error && resetErrors();
-
-  useEffect(() => {
-    if (isSuccess) {
-      setLoggedInUser(user);
-      push(routesPaths.index);
-    }
-  }, [isSuccess, user, push]);
-
-  useEffect(() => resetErrors, [resetErrors]);
+  const genders = ['male', 'female', 'other'];
+  const { onSubmit, isLoading, handleFocus, register, authenticated, error, errors } = useSignup({
+    genders,
+  });
 
   if (authenticated) {
     return <Redirect to={routesPaths.index} />;
@@ -65,7 +20,7 @@ const SignupForm = () => {
 
   return (
     <div className="signup">
-      <form onSubmit={handleSubmit(onSubmit)} noValidate>
+      <form onSubmit={onSubmit} noValidate>
         <div className="form-label">
           <label htmlFor="name">{t('signup.labels.name')}</label>
         </div>
@@ -112,8 +67,9 @@ const SignupForm = () => {
         </div>
         <Select
           name="gender"
-          options={['MALE', 'FEMALE', 'OTHER']}
+          options={genders}
           placeholder={t('signup.labels.genderSelect')}
+          error={errors.gender}
           handleFocus={handleFocus}
           register={register}
         />
